Simplify defaulting for system properties, and trust the environment less.
The originally-reported bug (https://android-review.googlesource.com/58258)
was that dalvik would fail to start if LD_LIBRARY_PATH was unset.

This changes things so we no longer have multiple default values for
java.library.path, cope better with that not being set, and don't set
it if LD_LIBRARY_PATH isn't set.

Change-Id: I55d48a35b1b42df872e6dcd0a38fc7bc4b8cad5a