NativeCrypto: free error state on client_cb
On a call to SSL_get_certificate() there appears to be one path that
will set an error code (SSL_aANON) and returns NULL. This client_cb
wasn't expecting NULL to be an error.

To avoid the error from lingering and showing up in later tests, clear
it so the later tests have more accurate errors.

Bug: 9198763
Change-Id: I229c06fce042436a8545812c1d48c1210408d57e