Fix parsing of fractional seconds.
We were just parsing them as millisecond values, which is wrong.
".1" fractional second is 100 milliseconds (we were parsing it as 1 ms),
and ".765432" fractional seconds is 765 milliseconds (we were
parsing it as 765432ms.

In most cases, this was hidden by the fact that we were formatting
such values incorrectly as well.

bug: 16969112
Change-Id: Id287684f78691090ac4a6c3029d7909f2c1e1310

Conflicts:
	luni/src/test/java/libcore/java/text/SimpleDateFormatTest.java